+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2025-04-30
+
+### Added
+- Sitemap API support with download functionality
+- Sitemap visualization endpoints (graph and markdown formats)
+- Tests for new sitemap API features
+
+### Changed
+- Simplified documentation with clearer examples
+- Updated LICENSE copyright to WaterCrawl-NodeJS
+
 ## [1.0.0] - 2025-01-19
 
 ### Added